oose-0 : -simplify coebase to use basic html and javascript - add opensearch.xml
