+++ opensearch.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<OpenSearchDescription xmlns="http://a9.com/-/spec/opensearch/1.1/">
+    <ShortName>Simple Unduck</ShortName>
+    <Description>Search using Simple Unduck</Description>
+    <InputEncoding>UTF-8</InputEncoding>
+    <Url type="text/html" template="http://localhost:63342/oose/?q={searchTerms}"/>
+    <Image height="16" width="16" type="image/x-icon">http://localhost:63342/oose/favicon.ico</Image>
+</OpenSearchDescription>
