+++ src/bang.js
