+++ tsconfig.json
@@ -1,4 +1,4 @@
-{
+
   "compilerOptions": {
     "target": "ES2020",
     "useDefineForClassFields": true,
